requirements.txt
alembic==1.7.5
APScheduler==3.6.3
backports.zoneinfo==0.2.1
borb==2.0.18
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fonttools==4.27.1
greenlet==1.1.2
gunicorn==20.1.0
icalendar==4.0.9
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
Pillow==8.4.0
python-barcode==0.13.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
qrcode==7.3.1
requests==2.27.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
style==1.1.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
update==0.0.1
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.6.0
